requirements.txt
pytest==7.4.2
Appium-Python-Client==3.0.0
